corregidos los permisos para el modulo convocatoria, y todos los modulos que estan relacionados con esta de manera directa

+++ apps/convocatorias/modules/documentos/templates/indexSuccess.php
@@ -1,8 +1,10 @@
 <h1>Plantillas de documentos requisitos</h1>
 
+<?php if ($sf_user->hasCredential('plantillas_create')): ?>
 <div class="tasks"><?php echo link_to(
     'Crear nuevo documento requisito', url_for('documentos_new')
 ) ?></div>
+<?php endif; ?>
 
 <table>
     <tr class="header">
@@ -14,15 +16,20 @@
         <td><?php echo $documento->getTexto() ?></td>
         <td>
             <ul>
+            <?php if ($sf_user->hasCredential('plantillas_edit')): ?>
                 <li><?php echo link_to(
                     'Editar', 'documentos_edit', $documento
                 ) ?></li>
+            <?php endif; ?>
+            <?php if ($sf_user->hasCredential('plantillas_delete')): ?>
                 <li><?php echo link_to(
                     'Eliminar', 'documentos_delete', $documento, array(
                         'method' => 'delete',
-                        'confirm' => '¿Esta seguro que desea eliminar el documento?'
+                        'confirm' => '¿Esta seguro que desea eliminar '
+                            . 'el documento?',
                     )
                 ) ?></li>
+            <?php endif; ?>
             </ul>
         </td>
     </tr>
