+++ apps/convocatorias/modules/requisitos/templates/indexSuccess.php
@@ -1,8 +1,10 @@
 <h1>Plantillas de requisitos</h1>
 
+<?php if ($sf_user->hasCredential('plantillas_create')): ?>
 <div class="tasks"><?php echo link_to(
     'Crear nuevo requisito', url_for('requisitos_new')
 ) ?></div>
+<?php endif; ?>
 
 <table>
     <tr class="header">
@@ -14,15 +16,21 @@
         <td><?php echo $requisito->getTexto() ?></td>
         <td>
             <ul>
+            <?php if ($sf_user->hasCredential('plantillas_edit')): ?>
                 <li><?php echo link_to(
                     'Editar', 'requisitos_edit', $requisito
                 ) ?></li>
+            <?php endif; ?>
+            <?php if ($sf_user->hasCredential('plantillas_delete')): ?>
                 <li><?php echo link_to(
-                    'Eliminar', 'requisitos_delete', $requisito, array(
+                    'Eliminar', 'requisitos_delete', $requisito,
+                    array(
                         'method' => 'delete',
-                        'confirm' => '¿Esta seguro que desea eliminar el requisito?'
+                        'confirm' => '¿Esta seguro que desea eliminar '
+                            . 'el requisito?',
                     )
                 ) ?></li>
+            <?php endif; ?>
             </ul>
         </td>
     </tr>
